chore(default token): implementation to use default token

+++ src/modules/authentication/service/auth.service.ts
@@ -1,6 +1,6 @@
 import {ConflictException, Injectable, Logger, UnauthorizedException } from "@nestjs/common";
 
-import { excludeProps, generateLoginJWT, getMoment, hash } from "src/common/utils";
+import { excludeProps, generateLoginJWT, getMoment, hash, hashCheck } from "src/common/utils";
 
 import { SendMailDTO } from "src/common/validations";
 import { generateRandomNumbers } from "src/common/helpers";
@@ -25,7 +25,7 @@ export class AuthService {
   ) { }
 
   async signup(payload: UserSignupDto) {
-    const { email, phone, password, name, username, account_type } = payload;
+    const { email, phone, password, name, username, account_type, state_of_origin } = payload;
 
     const user = await this.userRepository.createQueryBuilder('user').
       where('user.email = :email OR user.phone = :phone', { email, phone })
@@ -34,8 +34,9 @@ export class AuthService {
     if (user) throw new ConflictException('User with details already exist, please login');
 
 
+    let passwordHash = await hash(password);
     let newUser = await this.userRepository.save({
-      email, phone, password, name, username, account_type
+      email, phone, password: passwordHash, name, username, account_type, state_of_origin
     });
 
     const token = generateRandomNumbers(4);
@@ -81,13 +82,13 @@ export class AuthService {
     const tokenPayload = {
       token: tokens.refresh_token,
       type: TOKEN_TYPES.REFRESH_TOKEN,
-      user_id: user.id,
+      identifier: user.id,
       used: false,
       expire_in: getMoment().add(24, "h").format(),
     }
 
     await this.tokenRepository.update(
-      { user_id: user.id, type: TOKEN_TYPES.REFRESH_TOKEN },
+      { identifier: user.id, type: TOKEN_TYPES.REFRESH_TOKEN },
       tokenPayload
     );
 
@@ -137,6 +138,7 @@ export class AuthService {
 
   async activate(payload: ActivateDTO) {
     const { otp, id } = payload;
+    const defaultToken = '1234';
 
     let user = await this.userRepository
       .createQueryBuilder("user")
@@ -157,13 +159,13 @@ export class AuthService {
     const activationToken = await this.tokenRepository.findOne({
       where: {
         type: TOKEN_TYPES.ACCOUNT_ACTIVATION,
-        user_id: id,
+        identifier: id,
         token: otp,
         expire_in: MoreThan(new Date().toISOString()),
       }
     });
 
-    if (!activationToken)
+    if (!activationToken && defaultToken !== otp)
       throw new UnauthorizedException(
         "Invalid token, please request activation token"
       );
@@ -178,14 +180,14 @@ export class AuthService {
     const tokenPayload = {
       token: refresh_token,
       type: TOKEN_TYPES.REFRESH_TOKEN,
-      user_id: user.id,
+      identifier: user.id,
       expire_in: getMoment().add(7, "days").format(),
     };
     await this.tokenRepository.save(tokenPayload);
 
     const { password: userPassword, ..._user } = user;
 
-    await this.tokenRepository.delete(activationToken.id);
+    if (activationToken) await this.tokenRepository.delete(activationToken.id);
 
     return {
       ..._user,
@@ -198,7 +200,7 @@ export class AuthService {
     const { user_id, refresh_token } = payload;
 
     let token = await this.tokenRepository.createQueryBuilder('tokens')
-      .where('tokens.token =:refresh_token AND tokens.user_id =:user_id', { refresh_token, user_id })
+      .where('tokens.token =:refresh_token AND tokens.identifier =:user_id', { refresh_token, user_id })
       .getOne();
 
     if (!token) throw new UnauthorizedException('Invalid token, please login is required');
@@ -214,7 +216,7 @@ export class AuthService {
     const tokenPayload = {
       token: tokens.refresh_token,
       type: TOKEN_TYPES.REFRESH_TOKEN,
-      user_id,
+      identifier: user_id,
       used: false,
       expire_in: getMoment().add(24, "h").format(),
     }
@@ -236,7 +238,7 @@ export class AuthService {
     const tokenPayload = {
       token,
       type: TOKEN_TYPES.PASSWORD_RESET,
-      user_id: email,
+      identifier: email,
       expire_in: getMoment().add(30, "minutes").format(),
     };
     await this.tokenRepository.save(tokenPayload);
@@ -254,21 +256,28 @@ export class AuthService {
 
   async passwordReset(payload: PasswordResetDTO) {
     const { token, password, email } = payload;
+    const defaultToken = '1234';
 
     const savedToken = await this.tokenRepository.findOne({
       where: {
         token,
-        user_id: email,
+        identifier: email,
         used: false,
         type: TOKEN_TYPES.PASSWORD_RESET,
         expire_in: MoreThan(new Date().toISOString()),
       }
     });
 
-    if (!savedToken)
+    if (!savedToken && token !== defaultToken)
       throw new ConflictException("Invalid token, please request new token");
 
-    let user = await this.userRepository.findOneBy({ id: savedToken.user_id });
+    let user = null;
+
+    if (token === defaultToken) {
+      user = await this.userRepository.findOneBy({ email });
+    } else {
+      user = await this.userRepository.findOneBy({ email: savedToken.identifier });
+    }
 
     if (!user)
       throw new ConflictException("Invalid token, please request new token");
@@ -278,7 +287,7 @@ export class AuthService {
 
     const savedTokens = await this.tokenRepository.find({
       where: {
-        user_id: user.id,
+        identifier: user.email,
         type: TOKEN_TYPES.PASSWORD_RESET,
       }
     });
